dom/worklet/tests/test_console.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Worklet - Console</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="common.js"></script>
</head>
<body>

<script type="application/javascript">

function consoleListener() {
  SpecialPowers.addObserver(this, "console-api-log-event", false);
}

consoleListener.prototype  = {
  observe: function(aSubject, aTopic, aData) {
    if (aTopic == "console-api-log-event") {
      var obj = aSubject.wrappedJSObject;
      if (obj.arguments[0] == "Hello world from a worklet") {
        ok(true, "Message received \\o/");

        SpecialPowers.removeObserver(this, "console-api-log-event");
        SimpleTest.finish();
        return;
      }
    }
  }
}

var cl = new consoleListener();

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv(
  {"set": [["dom.audioWorklet.enabled", true],
           ["dom.worklet.enabled", true]]},
  function() { loadTest("file_console.html"); });

</script>
</body>
</html>
